Buyer

+++ apps/buyer-backend-spring/src/main/java/ccm/buyer/dto/response/BuyerResponse.java
@@ -1,12 +1,13 @@
 package ccm.buyer.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record BuyerResponse(
     Long id,
     String name,
     String email,
-    Double balance,
+    BigDecimal balance,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
 ) {}
