+++ apps/buyer-backend-spring/src/main/java/ccm/buyer/service/ListingService.java
@@ -1,11 +1,13 @@
 package ccm.buyer.service;
 
 import ccm.buyer.entity.Listing;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ListingService {
     List<Listing> getAll();
     Listing validateOpen(Long id);
-    void reserve(Long id, int qty);
-    void release(Long id, int qty);
+    void reserve(Long id, BigDecimal qty);
+    void release(Long id, BigDecimal qty);
 }
