+++ apps/buyer-backend-spring/src/main/java/ccm/buyer/service/PaymentService.java
@@ -1,7 +1,9 @@
 package ccm.buyer.service;
 
+import java.math.BigDecimal;
+
 import ccm.buyer.entity.Payment;
 
 public interface PaymentService {
-    Payment processPayment(Long trId, String method, Double amount);
+    Payment processPayment(Long trId, String method, BigDecimal amount);
 }
